# doc/offer-service-openapi.yaml
openapi: 3.0.0
info:
  title: Offer Service
  version: 1.0.0
  description: Lightning offer and metadata management service
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
paths:
...
  /offers/{partition}/{id}:
    get:
      summary: Get specific offer
      description: Retrieves detailed information about a specific payment offer including payment limits and metadata ID.
      parameters:
        - name: partition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Offer details
          headers:
            Cache-Control:
              schema:
                type: string
                example: "no-store, no-cache, must-revalidate"
            Expires:
              schema:
                type: string
                example: "Thu, 01 Jan 1970 00:00:00 GMT"
            Pragma:
              schema:
                type: string
                example: "no-cache"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferRecord'
        '404':
          description: Offer not found
    put:
      summary: Create or update offer
      description: Creates a new offer or updates an existing one with the provided payment configuration.
      parameters:
        - name: partition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferRecord'
      responses:
        '201':
          description: Offer created
        '204':
          description: Offer updated
    delete:
      summary: Remove offer
      description: Permanently removes a payment offer from the system.
      parameters:
        - name: partition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Offer removed
        '404':
          description: Offer not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OfferRecord:
      type: object
      required:
        - location
        - partition
        - id
        - maxSendable
        - minSendable
        - metadataId
        - timestamp
      properties:
        location:
          type: string
          description: Resource location path
          example: "default/123e4567-e89b-12d3-a456-426614174000"
        partition:
          type: string
        id:
          type: string
          format: uuid
        maxSendable:
          type: integer
          minimum: 0
        minSendable:
          type: integer
          minimum: 0
        metadataId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        expires:
          type: string
          format: date-time
...